feat(searchUtils): add readStream function
This commit adds a new readStream function to the searchUtils file. The function is used to read chunks of data from a ReadableStream and call a callback function with each chunk. The getResults function in searchUtils now uses this new readStream function instead of the previous implementation, which was removed.